salt: makepkg: added upstream_source & patches
The upstream_source argument uses yay to download the spec for the
specified package. It looks in the offical and aur repositories.

The patches argument is a list of file like (ex., salt:// URIs, ect)
items which are patches to be applied in order to the files in the
makepkg build directory (either from source or the new upstream_source argument).